Apirest/requirements.txt
asgiref
Django
djangorestframework
pytz
sqlparse
psycopg2
slugify
django-filter